changed type of real channels

+++ include/SimAnalogIn.hpp
@@ -14,8 +14,7 @@ namespace sim {
 		virtual double get();
 		
 	private:
-		
-		std::shared_ptr<sim::SimChannel<uint64_t>> chan;
+		std::shared_ptr<sim::SimChannel<double>> chan;
 	};
 };
